Refactored API logic out of the normal views.
