+++ okta_widget/views.py
@@ -9,7 +9,6 @@ from okta_widget.client.users_client import UsersClient
 from okta_widget.client.groups_client import GroupsClient
 from okta_widget.client.apps_client import AppsClient
 from okta_widget.forms import RegistrationForm, RegistrationForm2, TextForm, ActivationForm, ActivationWithEmailForm
-from okta_widget.decorators import access_token_required
 
 import json
 from django.contrib.auth.models import User
@@ -377,313 +376,6 @@ def view_admin(request):
     c.update({"impersonation_version": IMPERSONATION_VERSION})
     return render(request, 'admin.html', c)
 
-
-@access_token_required
-def list_users(request):
-    get = request.GET
-    startsWith = None
-    if 'startsWith' in get:
-        startsWith = get['startsWith']
-    client = UsersClient('https://' + OKTA_ORG, API_KEY)
-    profile = request.session['profile']
-    profile_dict = json.loads(profile)
-    companyName = profile_dict.get('companyName')
-
-    if 'admin' in request.session:
-        users = client.list_users(15, startsWith)
-    elif 'company_admin' in request.session:
-        users = client.list_users_scoped(15, companyName, startsWith)
-    else:
-        return not_authorized(request)
-
-    response = HttpResponse()
-    response.status_code = 200
-    response.content = users
-    return response
-
-
-@access_token_required
-def list_user(request):
-    get = request.GET
-    user_id = None
-    if 'user' in get:
-        user_id = get['user']
-    client = UsersClient('https://' + OKTA_ORG, API_KEY)
-
-    if 'admin' in request.session or 'company_admin' in request.session:
-        users = client.list_user(user_id)
-    else:
-        return not_authorized(request)
-
-    response = HttpResponse()
-    response.status_code = 200
-    response.content = users
-    return response
-
-
-@csrf_exempt
-@access_token_required
-def add_users(request):
-    response = HttpResponse()
-    response.status_code = 200
-
-    if request.method == 'POST':
-        req = request.POST
-
-        email = ''
-        firstName = ''
-        lastName = ''
-        role = ''
-        activate = False
-
-        profile_dict = json.loads(request.session['profile'])
-        companyName = ''
-        if 'companyName' in profile_dict:
-            companyName = profile_dict.get('companyName')
-
-        if 'email' in req:
-            email = req['email']
-        if 'firstName' in req:
-            firstName = req['firstName']
-        if 'lastName' in req:
-            lastName = req['lastName']
-        if 'role' in req:
-            role = req['role']
-        if 'activate' in req:
-            activate = req['activate']
-        client = UsersClient('https://' + OKTA_ORG, API_KEY)
-
-        user = {
-            "profile": {
-                "firstName": firstName,
-                "lastName": lastName,
-                "email": email,
-                "login": email,
-                "customer_role": role,
-                "companyName": companyName
-            }
-        }
-
-        if 'admin' in request.session:
-            users = client.create_user(user=user, activate=activate)
-        elif 'company_admin' in request.session:
-            users = client.create_user(user=user, activate=activate)
-            # users = client.create_user_scoped(user=user, activate="false", group="")
-        else:
-            return not_authorized(request)
-
-        response.content = users
-
-    return response
-
-
-@csrf_exempt
-@access_token_required
-def update_user(request):
-    response = HttpResponse()
-    response.status_code = 200
-
-    if request.method == 'POST':
-        req = request.POST
-
-        if 'user_id' in req:
-            user_id = req['user_id']
-
-            email = ''
-            firstName = ''
-            lastName = ''
-            role = ''
-            companyName = ''
-            deactivate = None
-
-            if 'email' in req:
-                email = req['email']
-            if 'firstName' in req:
-                firstName = req['firstName']
-            if 'lastName' in req:
-                lastName = req['lastName']
-            if 'role' in req:
-                role = req['role']
-            if 'deactivate' in req:
-                deactivate = req['deactivate']
-            if 'companyName' in req:
-                companyName = req['companyName']
-            client = UsersClient('https://' + OKTA_ORG, API_KEY)
-
-            user = {
-                "profile": {
-                    "firstName": firstName,
-                    "lastName": lastName,
-                    "email": email,
-                    "login": email,
-                    "customer_role": role,
-                    "companyName": companyName
-                }
-            }
-
-            if 'admin' in request.session:
-                users = client.update_user(user=user, user_id=user_id, deactivate=deactivate)
-            elif 'company_admin' in request.session:
-                users = client.update_user(user=user, user_id=user_id, deactivate=deactivate)
-            else:
-                return not_authorized(request)
-
-            response.content = users
-
-    return response
-
-
-@access_token_required
-def list_groups(request):
-    response = HttpResponse()
-    response.status_code = 200
-
-    profile = request.session['profile']
-    profile_dict = json.loads(profile)
-    companyName = ''
-    if 'companyName' in profile_dict:
-        companyName = profile_dict.get('companyName')
-
-    if 'company_admin' in request.session:
-        client = GroupsClient('https://' + OKTA_ORG, API_KEY)
-        response.content = client.list_groups(15, companyName)
-    else:
-        return not_authorized(request)
-
-    return response
-
-
-@access_token_required
-def get_group(request):
-    get = request.GET
-    response = HttpResponse()
-    response.status_code = 200
-
-    group_id = None
-    if 'group_id' in get:
-        group_id = get['group_id']
-    client = GroupsClient('https://' + OKTA_ORG, API_KEY)
-
-    if 'company_admin' in request.session:
-        response.content = client.get_group_by_id(group_id)
-    else:
-        return not_authorized(request)
-
-    return response
-
-
-@access_token_required
-def app_schema(request):
-    response = HttpResponse()
-    response.status_code = 200
-
-    if 'company_admin' in request.session:
-        client = AppsClient('https://' + OKTA_ORG, API_KEY, CLIENT_ID)
-        schema = client.get_schema()
-        response.content = schema
-    else:
-        return not_authorized(request)
-
-    return response
-
-
-@access_token_required
-def list_perms(request):
-    get = request.GET
-    response = HttpResponse()
-    response.status_code = 200
-
-    if 'company_admin' in request.session:
-        client = AppsClient('https://' + OKTA_ORG, API_KEY, CLIENT_ID)
-
-        group_id = None
-        if 'group_id' in get:
-            group_id = get['group_id']
-
-        perms = client.get_app_group_by_id(group_id)
-        response.content = perms
-    else:
-        return not_authorized(request)
-
-    return response
-
-
-@csrf_exempt
-@access_token_required
-def update_perm(request):
-    req = request.POST
-
-    group_id = None
-    perms = None
-
-    if 'group_id' in req:
-        group_id = req['group_id']
-    if 'perms' in req:
-        perms = req['perms']
-
-    response = HttpResponse()
-    response.status_code = 200
-
-    if 'company_admin' in request.session and group_id and group_id and perms:
-        if perms[-1:] == ',':
-            perms = perms[:-1]
-        perms = perms.split(',')
-        print(perms)
-
-        perm = {
-            "profile": {
-                "role_permissions": perms
-            }
-        }
-
-        client = AppsClient('https://' + OKTA_ORG, API_KEY, CLIENT_ID)
-        perms = client.update_app_group(group_id, perm)
-        response.content = perms
-    else:
-        return not_authorized(request)
-    return response
-
-
-@csrf_exempt
-@access_token_required
-def add_group(request):
-    response = HttpResponse()
-    response.status_code = 200
-
-    if request.method == 'POST':
-        req = request.POST
-        profile = request.session['profile']
-        profile_dict = json.loads(profile)
-
-        if 'groupName' in req and 'companyName' in profile_dict:
-            prefix = None
-            if 'companyName' in profile_dict:
-                prefix = profile_dict.get('companyName')
-                if prefix == '':
-                    prefix = None
-
-            group_name = req['groupName']
-            if prefix:
-                group_name = prefix + '_' + group_name
-
-            client = GroupsClient('https://' + OKTA_ORG, API_KEY)
-
-            group = {
-                "profile": {
-                    "name": group_name,
-                }
-            }
-
-            if 'admin' in request.session:
-                response.content = client.create_group(group)
-            elif 'company_admin' in request.session:
-                response.content = client.create_group(group)
-            else:
-                return not_authorized(request)
-
-    return response
-
-
 def view_debug(request):
     return render(request, 'debug.html', {'meta': request.META})
 
@@ -999,32 +691,6 @@ def login_delegate(request):
 
     c.update({"js": _do_format(request, '/js/login-delegate.js', 'login_delegate')})
     return render(request, 'login_delegate.html', c)
-
-
-# IMPERSONATION Demo
-@csrf_exempt
-@access_token_required
-def setNameId(request):
-    post = request.POST
-    print(post)
-
-    response = HttpResponse()
-    if 'nameid' in post:
-        version = '{}'.format(IMPERSONATION_VERSION)
-        if version == "1":
-            client = AppsClient('https://' + OKTA_ORG, API_KEY, IMPERSONATION_SAML_APP_ID)
-            response.status_code = client.set_name_id(request.session['user_id'], post['nameid'])
-        if version == "2":
-            u_client = UsersClient('https://' + IMPERSONATION_V2_ORG, IMPERSONATION_V2_ORG_API_KEY)
-            profile = request.session['profile']
-            users = u_client.list_user(json.loads(profile)['preferred_username'])
-            users = json.loads(users)
-            if "id" in users:
-                client = AppsClient('https://' + IMPERSONATION_V2_ORG, IMPERSONATION_V2_ORG_API_KEY, IMPERSONATION_V2_SAML_APP_ID)
-                response.status_code = client.set_name_id(users["id"], post['nameid'])
-                for key in list(request.session.keys()):
-                    del request.session[key]
-    return response
 
 
 @csrf_exempt
